Added dont@me feature

+++ src/utils/cache/database.js
@@ -1,6 +1,7 @@
 const { Schema, Model } = require("mongoose");
 const { redisClient } = require("../../Bot");
 const { debugMsg } = require("../TinyUtils");
+const env = require("../FetchEnvs")();
 
 const ONE_HOUR = 1 * 60 * 60; // Redis uses seconds.
 
@@ -13,8 +14,13 @@ class Database {
    *
    * @param {Schema} schema
    * @param {Map} model
+   * @param {boolean} [saveNull=false] - Optional, save null values to cache
    */
-  async findOne(schema, model) {
+  async findOne(schema, model, saveNull = false) {
+    var start = env.DEBUG_LOG ? Date.now() : undefined;
+    if (!schema || !model) {
+      throw new Error("Missing schema or model");
+    }
     const mongoKey = Object.keys(model)[0];
     const redisKey = schema.modelName + ":" + mongoKey + ":" + model[mongoKey];
     debugMsg(`Key: ${mongoKey} -> ${redisKey}`);
@@ -29,14 +35,16 @@ class Database {
       data = await schema.findOne(model);
       if (!data) {
         debugMsg(`Database miss no data found`);
-        return null;
+        if (!saveNull) return null;
       }
       await redisClient.set(redisKey, JSON.stringify(data));
       await redisClient.expire(redisKey, ONE_HOUR);
+      if (env.DEBUG_LOG) debugMsg(`Time taken: ${Date.now() - start}ms`);
       return data;
     }
 
     debugMsg(`Cache hit: ${redisKey} -> ${data}`);
+    if (env.DEBUG_LOG) debugMsg(`Time taken: ${Date.now() - start}ms`);
     return JSON.parse(data);
   }
 
@@ -56,6 +64,10 @@ class Database {
       new: true,
     }
   ) {
+    var start = env.DEBUG_LOG ? Date.now() : undefined;
+    if (!schema || !model) {
+      throw new Error("Missing schema or model");
+    }
     const mongoKey = Object.keys(model)[0];
     const redisKey = schema.modelName + ":" + mongoKey + ":" + model[mongoKey];
 
@@ -63,6 +75,7 @@ class Database {
     await redisClient.set(redisKey, JSON.stringify(object));
     await redisClient.expire(redisKey, ONE_HOUR);
 
+    if (env.DEBUG_LOG) debugMsg(`Time taken: ${Date.now() - start}ms`);
     debugMsg(`Updated key: ${mongoKey} -> ${redisKey}`);
   }
 
@@ -72,12 +85,17 @@ class Database {
    * @param {Map} model
    */
   async deleteOne(schema, model) {
+    var start = env.DEBUG_LOG ? Date.now() : undefined;
+    if (!schema || !model) {
+      throw new Error("Missing schema or model");
+    }
     const mongoKey = Object.keys(model)[0];
     const redisKey = schema.modelName + ":" + mongoKey + ":" + model[mongoKey];
     debugMsg(`Deleting key: ${mongoKey} -> ${redisKey}`);
 
     await redisClient.del(redisKey);
     await schema.deleteOne(model);
+    if (env.DEBUG_LOG) debugMsg(`Time taken: ${Date.now() - start}ms`);
   }
 }
